PII-25 removed abstract classes. Instead getting information from columns only to get lists of lines for rows and diagonals. Moved checking for wins out of column class and into board state class. Tidied code.
